Bugfix: Various calls to imagesAlign crashed due to wrong input data types.
Squashed commit of the following:

commit 68a9bf196bb51bddac6fd992ce28ee4dc69e8125
Author: Eric Kim <erickim555@gmail.com>
Date:   Wed Mar 13 18:25:54 2013 -0700

    Fixed a few dtype-related errors w.r.t the new imagesAlign Cython
    implementation (which assumes fixed types of its inputs, e.g.
    float32 vs. float64, float vs. None, etc).

+++ opencount/pixel_reg/doGrouping.py
@@ -205,7 +205,7 @@ def evalPatchSimilarity2(I,patch, debug=False):
 
     pixPad = round(.25 * min(patch.shape));
     patchPad = np.empty((patch.shape[0]+2*pixPad,
-                         patch.shape[1]+2*pixPad))
+                         patch.shape[1]+2*pixPad), dtype='float32')
     patchPad[:] = np.nan
     patchPad[pixPad:patch.shape[0]+pixPad,
              pixPad:patch.shape[1]+pixPad] = patch
